refactor: update button component and use it on index page
This commit refactors the button component to accept a `type` prop to specify the type of button as well as adding a `w-auto` class to the button's style. The index page also imports the `Button` component and `Input` component to render a form for users to submit their email. The styles for this form were adjusted accordingly.